backend/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
sqlalchemy
psycopg2-binary
